Move to io: verify_init_file, sign_init_file_metadata_and_poses and get_poses_from_init_file
