+++ source/src/python/PyRosetta/src/pyrosetta/distributed/cluster/init_files.py
@@ -9,44 +9,24 @@
 __author__ = "Jason C. Klima"
 
 
-try:
-    import toolz
-except ImportError:
-    print(
-        "Importing 'pyrosetta.distributed.cluster.init_files' requires the "
-        + "third-party package 'toolz' as a dependency!\n"
-        + "Please install this package into your python environment. "
-        + "For installation instructions, visit:\n"
-        + "https://pypi.org/project/toolz/\n"
-    )
-    raise
-
 import json
 import hmac
 import pyrosetta
 import pyrosetta.distributed.io as io
 
 from pyrosetta.distributed.packed_pose.core import PackedPose
-from pyrosetta.rosetta.basic import was_init_called
-from pyrosetta.rosetta.core.pose import get_all_comments
-from pyrosetta.utility.initialization import PyRosettaInitFileReader
 from typing import (
     Any,
     Dict,
     Generic,
     List,
-    NoReturn,
     Optional,
     Tuple,
     TypeVar,
-    Union,
 )
 
 from pyrosetta.distributed.cluster.hkdf import HASHMOD, compare_digest, derive_init_key
 
-
-METADATA_INPUT_DECOY_KEY: str = "idx_input"
-METADATA_OUTPUT_DECOY_KEY: str = "idx_output"
 
 G = TypeVar("G")
 
@@ -140,113 +120,3 @@ class InitFileSigner(Generic[G]):
     def verify(self, sha256: Optional[str], signature: Optional[str]) -> bool:
         """Verify PyRosetta initialization file SHA256 and HMAC signature metadata."""
         return self.verify_sha256(sha256) and self.verify_signature(signature)
-
-
-def verify_init_file(
-    init_file: str,
-    input_packed_pose: Optional[PackedPose],
-    output_packed_pose: Optional[PackedPose],
-    metadata: Dict[str, Any],
-) -> Optional[NoReturn]:
-    """Verify a PyRosetta initialization file."""
-
-    sha256 = metadata.pop("sha256", None)
-    signature = metadata.pop("signature", None)
-    signer = InitFileSigner(
-        input_packed_pose=input_packed_pose,
-        output_packed_pose=output_packed_pose,
-        metadata=metadata,
-    )
-    _err_msg = f"Could not verify data integrity of the input file: '{init_file}'"
-    if (sha256 is not None) and (not signer.verify_sha256(sha256)):
-        raise ValueError(
-            "The expected SHA256 value differs from the metadata 'sha256' key value in the '.init' file! "
-            + "The '.init' file 'poses' key value appears to have been altered from the original simulation "
-            + "or '.init' file export, so the simulation cannot necessarily be reproduced! To override "
-            + "data integrity verification, delete the 'sha256' key from the metadata in the '.init' file.\n"
-            + "Expected: '{0}'\n".format(signer.sign_sha256())
-            + "Value:    '{0}'\n".format(sha256)
-            + _err_msg
-        )
-    if (signature is not None) and (not signer.verify_signature(signature)):
-        raise ValueError(
-            "The expected PyRosettaCluster signature differs from the metadata 'signature' key value in the "
-            + "'.init' file! The '.init' file data appears to have been altered from the original simulation "
-            + "or '.init' file export, so the simulation cannot necessarily be reproduced! To override "
-            + "data integrity verification, delete the 'signature' key from the metadata in the '.init' file.\n"
-            + "Expected: '{0}'\n".format(signer.sign_digest())
-            + "Value:    '{0}'\n".format(signature)
-            + _err_msg
-        )
-    if all(val is not None for val in (sha256, signature)):
-        assert signer.verify(sha256, signature), _err_msg
-
-
-def sign_init_file_metadata_and_poses(
-    input_packed_pose: Optional[PackedPose] = None,
-    output_packed_pose: Optional[PackedPose] = None,
-) -> Tuple[Dict[str, Any], List[PackedPose]]:
-    """Sign PyRosetta initialization file 'metadata' and 'poses' keys."""
-
-    metadata = {}
-    metadata["comment"] = "Generated by PyRosettaCluster"
-    metadata["version"] = pyrosetta.distributed.cluster.__version__
-    poses = []
-    if isinstance(output_packed_pose, PackedPose): # Set output decoy to front of poses list if present
-        _comments = dict(get_all_comments(output_packed_pose.pose))
-        if "REMARK PyRosettaCluster:" not in _comments.keys():
-            raise NotImplementedError(
-                "Signing '.init' file metadata and poses when PyRosettaCluster simulation data is not "
-                + f"cached in the output `PackedPose` object comments is not supported: {_comments}"
-            )
-        poses.append(output_packed_pose)
-        metadata[METADATA_OUTPUT_DECOY_KEY] = poses.index(output_packed_pose)
-    if isinstance(input_packed_pose, PackedPose):
-        poses.append(input_packed_pose)
-        metadata[METADATA_INPUT_DECOY_KEY] = poses.index(input_packed_pose)
-    signer = InitFileSigner(
-        input_packed_pose=input_packed_pose,
-        output_packed_pose=output_packed_pose,
-        metadata=metadata,
-    )
-    metadata.update(signer.sign())
-
-    return metadata, poses
-
-
-def get_poses_from_init_file(
-    init_file: str,
-    verify: bool = False,
-) -> Union[Tuple[Optional[PackedPose], Optional[PackedPose]], NoReturn]:
-    """
-    Return a `tuple` object of the input `PackedPose` object and the output `PackedPose` object
-    from a '.init' file, and optionally verify PyRosettaCluster metadata in the '.init' file.
-    """
-
-    @toolz.functoolz.curry
-    def _maybe_to_packed(
-        key: str, poses: List[str], metadata: Dict[str, str]
-    ) -> Union[Optional[PackedPose], NoReturn]:
-        assert was_init_called(), (
-            f"Please first initialize PyRosetta with the 'init_file' argument parameter: '{init_file}'"
-        )
-        assert isinstance(metadata, dict), (
-            f"The '.init' file 'metadata' key value must be a `dict` object. Received: {type(metadata)}"
-        )
-        assert isinstance(poses, list), (
-            f"The '.init' file 'poses' key value must be a `list` object. Received: {type(poses)}"
-        )
-
-        return io.to_packed(io.to_pose(poses[metadata[key]])) if key in metadata else None
-
-    init_dict = PyRosettaInitFileReader.read_json(init_file)
-    metadata = init_dict["metadata"]
-    poses = init_dict["poses"]
-    _get_packed_pose = _maybe_to_packed(poses=poses, metadata=metadata)
-    input_packed_pose = _get_packed_pose(METADATA_INPUT_DECOY_KEY)
-    output_packed_pose = _get_packed_pose(METADATA_OUTPUT_DECOY_KEY)
-
-    if verify:
-        verify_init_file(init_file, input_packed_pose, output_packed_pose, metadata)
-
-    return (input_packed_pose, output_packed_pose)
